+++ src/test/java/org/graylog/plugins/netflow/codecs/NetFlowCodecTest.java
@@ -2,9 +2,6 @@ package org.graylog.plugins.netflow.codecs;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Resources;
-import io.pkts.Pcap;
-import io.pkts.packet.UDPPacket;
-import io.pkts.protocol.Protocol;
 import org.graylog.plugins.netflow.flows.FlowException;
 import org.graylog2.plugin.Message;
 import org.graylog2.plugin.configuration.Configuration;
@@ -16,14 +13,11 @@ import org.junit.rules.TemporaryFolder;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.InputStream;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
@@ -132,191 +126,5 @@ public class NetFlowCodecTest {
         final InetSocketAddress source = new InetSocketAddress(InetAddress.getLocalHost(), 12345);
 
         assertThat(codec.decodeMessages(new RawMessage(b, source))).isNull();
-    }
-
-    @Test
-    public void pcap_softflowd_NetFlowV5() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/netflow5.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages)
-                                    .isNotNull()
-                                    .isNotEmpty();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-            assertThat(allMessages)
-                    .hasSize(4)
-                    .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(5));
-        }
-    }
-
-    @Test
-    public void pcap_pmacctd_NetFlowV5() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/pmacctd-netflow5.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages)
-                                    .isNotNull()
-                                    .isNotEmpty();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-            assertThat(allMessages)
-                    .hasSize(42)
-                    .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(5));;
-        }
-    }
-
-    @Test
-    public void pcap_softflowd_NetFlowV9() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/netflow9.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages)
-                                    .isNotNull()
-                                    .isNotEmpty();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-        }
-        assertThat(allMessages)
-                .hasSize(19)
-                .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(9));;
-    }
-
-    @Test
-    public void pcap_pmacctd_NetFlowV9() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/pmacctd-netflow9.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages)
-                                    .isNotNull()
-                                    .isNotEmpty();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-        }
-        assertThat(allMessages)
-                .hasSize(6)
-                .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(9));;
-    }
-
-    @Test
-    public void pcap_netgraph_NetFlowV5() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/netgraph-netflow5.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages)
-                                    .isNotNull()
-                                    .isNotEmpty();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-        }
-        assertThat(allMessages)
-                .hasSize(120)
-                .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(5));
-    }
-
-    @Test
-    public void pcap_nprobe_NetFlowV9_mixed() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/nprobe-netflow9.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages).isNotNull();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-        }
-        assertThat(allMessages)
-                .hasSize(152);
-    }
-
-    @Test
-    public void pcap_nprobe_NetFlowV9_2() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/nprobe-netflow9-2.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            assertThat(messages).isNotNull();
-                            allMessages.addAll(messages);
-                        }
-                        return true;
-                    }
-            );
-        }
-        assertThat(allMessages)
-                .hasSize(6)
-                .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(9));
-    }
-
-    @Test
-    public void pcap_nprobe_NetFlowV9_4() throws Exception {
-        final List<Message> allMessages = new ArrayList<>();
-        try (InputStream inputStream = Resources.getResource("netflow-data/nprobe-netflow9-4.pcap").openStream()) {
-            final Pcap pcap = Pcap.openStream(inputStream);
-            pcap.loop(packet -> {
-                        if (packet.hasProtocol(Protocol.UDP)) {
-                            final UDPPacket udp = (UDPPacket) packet.getPacket(Protocol.UDP);
-                            final InetSocketAddress source = new InetSocketAddress(udp.getSourceIP(), udp.getSourcePort());
-                            final Collection<Message> messages = codec.decodeMessages(new RawMessage(udp.getPayload().getArray(), source));
-                            if (messages != null) {
-                                allMessages.addAll(messages);
-                            }
-                        }
-                        return true;
-                    }
-            );
-        }
-        assertThat(allMessages)
-                .hasSize(1)
-                .allSatisfy(message -> assertThat(message.getField("nf_version")).isEqualTo(9));
     }
 }
